+++ src/fixate/drivers/dmm/fluke_8846a.py
@@ -57,7 +57,7 @@ class Fluke8846A(DMM):
 
         # High and low current port definition. Each definition encodes the maximum current able to
         # be measured by the port (in amps)
-        self.current_ports = {"HIGH": 10, "LOW": 400e-3}
+        self._current_ports = {"HIGH": 10, "LOW": 400e-3}
 
     @property
     def samples(self):
@@ -284,17 +284,8 @@ class Fluke8846A(DMM):
         exception is raised.
         """
 
-        # Check the requested range is not more than the port capability:
-        if _range > self.current_ports[port]:
-            raise ValueError(
-                "The selected port and range combination is not available for this instrument. Consider using a different multimeter"
-            )
-
-        # Raise an error if the high port is selected when the low port should be used:
-        if _range < self.current_ports["LOW"] and port == "HIGH":
-            raise ValueError(
-                "High range port selected when the low range port should be used! Consider using a different multimeter."
-            )
+        # Check the port and range combination is compatible for the instrument:
+        self._check_current_port_range(_range, port)
 
         self._set_measurement_mode("current_ac", _range)
 
@@ -309,17 +300,8 @@ class Fluke8846A(DMM):
         exception is raised.
         """
 
-        # Check the requested range is not more than the port capability:
-        if _range > self.current_ports[port]:
-            raise ValueError(
-                "The selected port and range combination is not available for this instrument. Consider using a different multimeter"
-            )
-
-        # Raise an error if the high port is selected when the low port should be used:
-        if _range < self.current_ports["LOW"] and port == "HIGH":
-            raise ValueError(
-                "High range port selected when the low range port should be used! Consider using a different multimeter."
-            )
+        # Check the port and range combination is compatible for the instrument:
+        self._check_current_port_range(_range, port)
 
         self._set_measurement_mode("current_dc", _range)